+++ NEWS.md
@@ -1,4 +1,4 @@
-# mlr3benchmark 0.1.0.9000
+# mlr3benchmark 0.1.1
 
 * BenchmarkAggr$friedman_test now returns the full test object if only a single measure exists in the object
 * Fixed plotting in `autoplot.BenchmarkAggr` for CD-plots, previously bars were overlapping and giving misleading results.